refactor: Summary 부모 엔티티 생성 및 AI 생성 요약, 사용자 생성 요약이 이를 상속
